+++ src/main/java/org/sparta/tech259/finalproject/model/exception/comments/CommentBodyNotFoundException.java
@@ -1,4 +1,4 @@
-package org.sparta.tech259.finalproject.model.exceptions;
+package org.sparta.tech259.finalproject.model.exception.comments;
 
 public class CommentBodyNotFoundException extends Exception {
      public CommentBodyNotFoundException() {
