+++ src/main/java/org/sparta/tech259/finalproject/model/exception/comments/Response.java
@@ -1,4 +1,4 @@
-package org.sparta.tech259.finalproject.model.exceptions;
+package org.sparta.tech259.finalproject.model.exception.comments;
 
 public record Response(int statusCode, String url, String message) {
 }
